# api/openapi/swagger.yaml
components:
  schemas:
    errs.Error:
      properties:
        code:
          $ref: '#/components/schemas/errs.ErrorCode'
        message:
          type: string
        params:
          items:
            $ref: '#/components/schemas/errs.Param'
          type: array
          uniqueItems: false
      type: object
    errs.ErrorCode:
      type: integer
      x-enum-varnames:
      - ErrorCodeOK
      - ErrorCodeCanceled
      - ErrorCodeUnknown
      - ErrorCodeInvalidArgument
      - ErrorCodeDeadlineExceeded
      - ErrorCodeNotFound
      - ErrorCodeAlreadyExists
      - ErrorCodePermissionDenied
      - ErrorCodeResourceExhausted
      - ErrorCodeFailedPrecondition
      - ErrorCodeAborted
      - ErrorCodeOutOfRange
      - ErrorCodeUnimplemented
      - ErrorCodeInternal
      - ErrorCodeUnavailable
      - ErrorCodeDataLoss
      - ErrorCodeUnauthenticated
      - ErrorCodeClosedRequest
    errs.Param:
      properties:
        key:
          type: string
        value:
          type: string
      type: object
...
    handlers.LikeCreateDTO:
      properties:
        post_id:
          type: string
        user_id:
          type: string
        value:
          type: string
      type: object
    handlers.LikeDTO:
      properties:
        created_at:
          type: string
        id:
          type: string
        post_id:
          type: string
        updated_at:
          type: string
        user_id:
          type: string
        value:
          type: string
      type: object
    handlers.LikeListDTO:
      properties:
        count:
          type: integer
        items:
          items:
            $ref: '#/components/schemas/handlers.LikeDTO'
          type: array
          uniqueItems: false
      type: object
    handlers.LikeUpdateDTO:
      properties:
        id:
          type: string
        post_id:
          type: string
        user_id:
          type: string
        value:
          type: string
      type: object
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  title: example
  version: 0.0.0
openapi: 3.1.0
paths:
...
  /api/v1/posts/likes/:
    get:
      parameters:
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: page_number
        schema:
          type: integer
      - in: query
        name: order_by
        schema:
          items:
            type: string
          type: array
          uniqueItems: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.LikeListDTO'
          description: Filtered list of likes
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: List of likes
      tags:
      - like
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.LikeCreateDTO'
        description: Create like request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.LikeDTO'
          description: Created like
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Create like
      tags:
      - like
  /api/v1/posts/likes/{id}:
    delete:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete like by id
      tags:
      - like
    get:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.LikeDTO'
          description: Requested like
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get like by id
      tags:
      - like
    patch:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.LikeUpdateDTO'
        description: Update like request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.LikeDTO'
          description: Updated like
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Update like
      tags:
      - like
...
servers:
- url: http://127.0.0.1:8000/